+++ Online/State/RealizedSpearState.cs
@@ -56,8 +56,6 @@ namespace RainMeadow
             spear.stuckInWall = stuckInWall;
             spear.abstractSpear.stuckInWallCycles = stuckInWallCycles;
             spear.spearDamageBonus = spearDamageBonus;
-            spear.addPoles = stuckInWall.HasValue;
-
             spear.spearmasterNeedle_hasConnection = needleActive;
 
             if (stuckInChunk is not null)
@@ -69,11 +67,7 @@ namespace RainMeadow
                 spear.stuckRotation = stuckRotation;
             }
 
-            if (spear is ExplosiveSpear explosive) {
-                if (ignited && !explosive.Ignited) {
-                    explosive.Ignite();
-                }
-            }
+            if (spear is ExplosiveSpear explosive && ignited && !explosive.Ignited) explosive.Ignite();
 
             base.ReadTo(onlineEntity);
             if (spear.mode == Weapon.Mode.StuckInWall && !spear.stuckInWall.HasValue)
@@ -91,9 +85,9 @@ namespace RainMeadow
 
         override public bool ShouldPosBeLenient(PhysicalObject po)
         {
-            if (po is not Spear p) { RainMeadow.Error("target is wrong type: " + po); return false; }
+            if (po is not Spear p) { RainMeadow.Error("target is wrong type: " + po); return base.ShouldPosBeLenient(po); }
             if (p.onPlayerBack) return true;
-            if (p.stuckInObject != null) return true; 
+            if (p.mode == Weapon.Mode.StuckInCreature || p.mode == Weapon.Mode.StuckInWall) return true; 
             return base.ShouldPosBeLenient(po);
         }
     }